Using std::shared_ptr to manage VideoCapture and AusioCapture instances Merged RecordingOptions and EncoderOptions Fixed Win32DeviceWatcher to notify of devices changed Sockets now accept uv_event_loop constructor argument Pacman InstallTask is now event based (wasd thread based) Updated Pacman samples Fied typo in Symple Client signal IPC classes now optionally event based

+++ src/media/src/audiocontext.cpp
@@ -582,7 +582,7 @@ UInt8* AudioResampler::resample(UInt8* inSamples, int inNbSamples)
         throw std::runtime_error("Conversion context must be initialized.");	
 
 	outNbSamples = av_rescale_rnd(
-		swr_get_delay(ctx, (int64_t)iparams.sampleRate) + inNbSamples, 
+		swr_get_delay(ctx, (int64_t)iparams.sampleRate) + (int64_t)inNbSamples, 
 		(int64_t)oparams.sampleRate, (int64_t)iparams.sampleRate, AV_ROUND_UP);
 
 	if (outBuffer) {
